add abstract BaseStringTermsTestCase
